Add Hypothesis property-based tests, fix 3 bugs found (#555)
* Add Hypothesis property-based tests, uncovering 3 bugs

Adds ddmc/tests/test_properties.py with property-based tests (via
Hypothesis) covering the core sequence-distance and preprocessing
helpers in ddmc.pam250, ddmc.binomial, ddmc.motifs, and
ddmc.logistic_regression.

These tests found three real bugs, kept as strict xfail so CI stays
green while documenting them (and will loudly flag a regression if the
underlying bug is fixed without updating the test):

- get_pam250_scores accumulates pairwise scores in an int8 array, which
  silently overflows for highly self-similar peptides (e.g. poly-W or
  poly-C), producing wrong (sometimes negative) similarity scores.
- compute_control_pssm crashes with ValueError on a lowercase
  phosphoacceptor character, which is exactly what its real caller
  chain produces: BackgroundSeqs lowercases the phosphoacceptor, and
  DDMC.get_pssms(PsP_background=True) passes that output straight in.
- normalize_cluster_centers centers along the wrong axis: the
  docstring says it zero-means each cluster across samples, but the
  double-transpose around StandardScaler actually zero-means each
  sample across clusters instead.

The remaining property tests pass and pin down invariants (PWM
symmetry/equivalence to the Biopython reference, one-hot encoding
round-trips, phosphosite-type counting) that should keep holding as
this code changes.


* Fix the 3 bugs found by the new property-based tests

- get_pam250_scores: widen the pairwise-score accumulator from int8 to
  int32 so it no longer silently overflows (and wraps to a negative
  score) for highly self-similar peptides such as poly-tryptophan or
  poly-cysteine motifs.
- compute_control_pssm: upper-case each residue before the AAlist
  lookup, so it no longer crashes on the lowercased phosphoacceptor
  that BackgroundSeqs (its actual caller, via
  DDMC.get_pssms(PsP_background=True)) always produces.
- normalize_cluster_centers: drop the erroneous transpose-around-
  StandardScaler, which was centering each sample across its clusters
  instead of centering each cluster across samples as documented.

Turns the corresponding xfail(strict=True) tests in
ddmc/tests/test_properties.py into ordinary passing regression tests.


---------

+++ ddmc/pam250.py
@@ -72,7 +72,10 @@ def get_pam250_scores(seqs: list[str]) -> np.ndarray:
     # convert to np array
     pam250m = np.array(pam250.values(), dtype=np.int8).reshape(pam250.shape)
 
-    out = np.zeros((seq_idx.shape[0], seq_idx.shape[0]), dtype=np.int8)
+    # int32, not int8: an int8 accumulator overflows for realistic peptide
+    # lengths (e.g. 11 tryptophans score 11 * 17 = 187, above int8's 127
+    # ceiling), silently wrapping self-similarity scores to negative values.
+    out = np.zeros((seq_idx.shape[0], seq_idx.shape[0]), dtype=np.int32)
     i_idx, j_idx = np.tril_indices(seq_idx.shape[0])
     out[i_idx, j_idx] = np.sum(pam250m[seq_idx[i_idx], seq_idx[j_idx]], axis=1)
